update ability added to djanguify

+++ djangui/scripts/djanguify.py
@@ -24,7 +24,7 @@ def main():
     parser.add_argument('-r', '--recursive', help='Recursively search for scripts.', action='store_true')
     parser.add_argument('-a', '--app', help='The name of the app to create.', type=str, required=True)
     parser.add_argument('-p', '--project', help='The name of the django project to create.', type=str)
-    parser.add_argument('--fake', help='Dont do anything.', action='store_true')
+    parser.add_argument('--update', help='Update the model definitions of a given app.', action='store_true')
     args = parser.parse_args()
     scripts_path = args.scripts
     scripts = []
@@ -40,21 +40,21 @@ def main():
     project_name = args.project
     app_name = args.app
     app_base_dir = ''
+    update_app = args.update
     if project_name:
-        if not args.fake:
+        if not update_app:
             subprocess.call(['django-admin.py', 'startproject', project_name])
         project_root = os.path.join(app_base_dir, project_name)
         project_base_dir = os.path.join(project_root, project_name)
         app_base_dir = os.path.join(app_base_dir, project_name)
     app_base_dir = os.path.join(app_base_dir, app_name)
 
-    if not args.fake:
+    if not update_app:
         com = ['django-admin.py', 'startapp', app_name]
         if project_name:
             app_path = os.path.join(os.path.abspath(project_name), app_name)
             os.mkdir(app_path)
             com.append(app_path)
-        print com
         subprocess.call(com)
 
     djanguify_folder = os.path.split(os.path.realpath(__file__))[0]
@@ -102,21 +102,24 @@ def main():
     )
 
     template_files = []
-    def walk_dir(templates, dest):
+    def walk_dir(templates, dest, filter=None):
         l = []
         for root, folders, files in os.walk(templates):
             for filename in files:
-                if filename.endswith('.pyc'):
+                if filename.endswith('.pyc') or (filter and filename not in filter):
                     continue
                 relative_dir = '.{0}'.format(os.path.split(os.path.join(root, filename).replace(templates, ''))[0])
                 l.append((os.path.join(root, filename), os.path.join(dest, relative_dir)))
         return l
 
-    template_files += walk_dir(app_template_dir, app_base_dir)
-    template_files += walk_dir(project_template_dir, project_base_dir)
+    template_files += walk_dir(app_template_dir, app_base_dir, filter={'djangui_models.py'} if update_app else None)
+
+    if project_name and not update_app:
+        template_files += walk_dir(project_template_dir, project_base_dir)
 
     # remove files of directories we are overriding
-    os.remove(os.path.join(app_base_dir, 'models.py'))
+    if not update_app:
+        os.remove(os.path.join(app_base_dir, 'models.py'))
 
     for template_file, dest_dir in template_files:
         template_file = open(template_file)
@@ -135,10 +138,11 @@ def main():
         with open(to_name, 'wb') as new_file:
             new_file.write(content)
 
-    # move the django settings to the settings path so we don't have to chase Django changes.
-    shutil.move(os.path.join(project_base_dir, 'settings.py'), os.path.join(project_base_dir, 'settings', 'django_settings.py'))
-    # do the same with urls
-    shutil.move(os.path.join(project_base_dir, 'urls.py'), os.path.join(project_base_dir, 'urls', 'django_urls.py'))
+    if not update_app:
+        # move the django settings to the settings path so we don't have to chase Django changes.
+        shutil.move(os.path.join(project_base_dir, 'settings.py'), os.path.join(project_base_dir, 'settings', 'django_settings.py'))
+        # do the same with urls
+        shutil.move(os.path.join(project_base_dir, 'urls.py'), os.path.join(project_base_dir, 'urls', 'django_urls.py'))
 
     if project_name:
         subprocess.call(['python', os.path.join(project_root, 'manage.py'), 'makemigrations'])
